2025.05.02

+++ src/main/java/com/shTest/equipment/repository/EquipmentRepository.java
@@ -1,8 +1,7 @@
 package com.shTest.equipment.repository;
 
 import com.shTest.entity.Equipment;
-import com.shTest.equipment.dto.EquipmentDto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EquipmentRepository extends JpaRepository<Equipment, Long>, EquipmentCustomRepository {
+public interface EquipmentRepository extends JpaRepository<Equipment, Integer>, EquipmentCustomRepository {
 }
